# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IKINCI KANAL - Decentralized Secure Messenger API
  version: 0.2.0
  description: |
    Decentralized, end-to-end encrypted messaging platform.

    ## Core Principles
    - **Zero Knowledge**: Server never sees plaintext content
    - **Metadata Protection**: Sealed sender, minimal logging
    - **User Control**: One-click data deletion
    - **Decentralized**: Federated server architecture
    - **Disappearing Messages**: Auto-delete support

    ## Security Roadmap
    | Version | Authentication | Notes |
    |---------|---------------|-------|
    | **V1 (MVP)** | Passkey (WebAuthn) + password fallback | Passkey preferred, password for legacy |
    | **V2** | OPAQUE protocol | Password users: server NEVER sees password |
    | **V3** | OPAQUE + Hardware key | YubiKey/Titan key binding |

    **V1 Auth Strategy:**
    - **Primary**: Passkey (WebAuthn) - biometric/device-bound, phishing-resistant
    - **Fallback**: Password (Argon2id) - for devices without passkey support
    - **Recommendation**: Prompt users to add passkey even if registered with password

    **Why Passkey in V1?**
    - Mobile: Face ID / Touch ID = zero friction + maximum security
    - Desktop: Platform authenticator or security key
    - No password = no password to steal, leak, or phish

    **Why OPAQUE in V2?**
    For password fallback users: OPAQUE ensures server NEVER sees the password,
    only stores a "password file" useless without client computation.

...
    ```
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new account + first device
      security: []
      description: |
        Creates a new user account with initial device and keys.
        Handle format: alphanumeric, 3-32 chars.
        For federated users, handle becomes `handle@this-server.com`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with password (fallback)
      security: []
      description: |
        Password-based login. For passkey login, use /auth/passkey/* endpoints.

        **Rate Limiting (Brute-Force Protection):**
        | Scope | Limit | Window | Action |
        |-------|-------|--------|--------|
        | Per IP | 5 attempts | 15 min | 429 + Retry-After |
        | Per handle | 10 attempts | 1 hour | 429 + exponential backoff |
        | Global suspicious | 100/min | 1 min | CAPTCHA required |

        After 10 failed attempts per handle, account enters "cool-down" mode:
        - 1st lockout: 15 min
        - 2nd lockout: 1 hour
        - 3rd lockout: 24 hours + email/passkey verification required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/passkey/options:
    post:
      tags: [Auth]
      summary: Get WebAuthn challenge for passkey login
      security: []
      description: |
        Step 1 of passkey login. Returns challenge for navigator.credentials.get().

        **Flow:**
        1. Client calls this endpoint with handle
        2. Server returns challenge + allowCredentials (user's registered passkeys)
        3. Client calls navigator.credentials.get() with options
        4. Client sends assertion to /auth/passkey/verify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                handle:
                  type: string
                  description: User handle to login
              required: [handle]
      responses:
        "200":
          description: WebAuthn options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyLoginOptions"
        "404":
          description: User not found or no passkey registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/passkey/verify:
    post:
      tags: [Auth]
      summary: Verify passkey assertion and login
      security: []
      description: |
        Step 2 of passkey login. Verifies assertion from navigator.credentials.get().
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                handle:
                  type: string
                assertion:
                  $ref: "#/components/schemas/PasskeyAssertion"
                device:
                  $ref: "#/components/schemas/DeviceBootstrap"
                  description: Device info for new device login
                deviceKeys:
                  $ref: "#/components/schemas/InitialKeys"
                  description: Keys for new device (if new device)
              required: [handle, assertion]
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid assertion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      description: |
        Invalidates the current session for this device.
        Revokes the refresh token (access token expires naturally or via jti blacklist).
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # USER PROFILE & DISCOVERY
  # ============================================================
...
  /devices:
    get:
      tags: [Devices]
      summary: List my devices
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
                required: [devices]
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Devices]
      summary: Add new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDeviceRequest"
      responses:
        "201":
          description: Device added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: Revoke/logout a device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "204":
          description: Device revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /devices/revoke-all:
    post:
      tags: [Devices, Privacy]
      summary: Logout from ALL devices
      description: |
        Revokes all devices except current one.
        Use when you suspect unauthorized access.
      responses:
        "200":
          description: All devices revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revokedCount:
                    type: integer
                    example: 3
                required: [revokedCount]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /devices/revoke-all-including-current:
    post:
      tags: [Devices, Privacy]
      summary: Logout from ALL devices including current
      description: Emergency logout from everywhere
      responses:
        "204":
          description: All devices revoked, session ended
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # KEY MANAGEMENT (Signal Protocol)
  # ============================================================
  /keys/prekey-bundle/{userId}:
    get:
      tags: [Keys]
      summary: Get prekey bundle for session establishment
      description: |
        Returns prekey bundle for initiating encrypted session.
        For federated users, proxies request to their home server.
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Specific device (optional, server picks primary if omitted)
      responses:
        "200":
          description: Prekey bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreKeyBundle"
        "404":
          $ref: "#/components/responses/NotFound"
  /keys/upload:
    post:
      tags: [Keys]
      summary: Upload new prekeys
      description: |
        Upload signed prekey and one-time prekeys.
        Should be called when one-time prekeys are running low.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadKeysRequest"
      responses:
        "200":
          description: Keys uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  remainingOneTimeKeys:
                    type: integer
                    description: How many one-time keys left on server
                  serverTime:
                    type: string
                    format: date-time
                required: [ok, remainingOneTimeKeys, serverTime]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/count:
    get:
      tags: [Keys]
      summary: Get remaining one-time prekey count
      description: Check if you need to upload more prekeys
      responses:
        "200":
          description: Key count
          content:
            application/json:
              schema:
                type: object
                properties:
                  oneTimeKeyCount:
                    type: integer
                    example: 42
                  recommendedMinimum:
                    type: integer
                    example: 20
                required: [oneTimeKeyCount, recommendedMinimum]
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # MESSAGING (with Sealed Sender & Disappearing Messages)
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User authentication token

    federationAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        Server-to-server authentication via Ed25519 request signing.

        **Required Headers:**
        - `X-Server-Origin`: Source server URL (e.g., https://serverA.com)
        - `X-Server-Key-Id`: Public key identifier for rotation (e.g., key-2025-01)
        - `X-Timestamp`: ISO 8601 timestamp (max ±5 min skew)
        - `X-Nonce`: Random 16+ char string (replay protection)
        - `X-Signature`: Base64 Ed25519 signature

        **Signature Computation:**
        ```
        canonical = HTTP_METHOD + "\n"
                  + REQUEST_PATH + "\n"
                  + X-Timestamp + "\n"
                  + X-Nonce + "\n"
                  + SHA256_HEX(request_body)

        signature = Base64(Ed25519.sign(canonical, server_private_key))
        ```

        **Verification:**
        1. Check X-Timestamp within ±5 minutes
        2. Check X-Nonce not seen from this origin in last 5 min (replay)
        3. Fetch sender's public key from `/.well-known/ikincikanal-server`
        4. Verify signature matches canonical request
        5. Cache (origin, nonce) for 5 min

        **Key Rotation:**
        - Rotate keys monthly (recommended)
...
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UNAUTHORIZED
            message: Missing or invalid token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: NOT_FOUND
            message: Resource not found
    Conflict:
      description: Conflict (duplicate)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: CONFLICT
            message: Already exists
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    ValidationError:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required: [code, message, fields]

    # ==================== HEALTH ====================
...
    RegisterRequest:
      type: object
      description: |
        Register with EITHER password OR passkey (not both).
        Passkey is preferred for security.
      properties:
        handle:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_\\.]+$"
          example: fatih
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: |
            Password auth (fallback). Min 12 chars.
            Storage: Argon2id (time=3, mem=64MB, parallel=4).
            Mutually exclusive with passkey.
        passkey:
          $ref: "#/components/schemas/PasskeyAttestation"
          description: |
            **Preferred.** WebAuthn attestation from navigator.credentials.create().
            Mutually exclusive with password.
        device:
          $ref: "#/components/schemas/DeviceBootstrap"
        keys:
          $ref: "#/components/schemas/InitialKeys"
        acceptedPrivacyVersion:
          type: string
          example: "2025-01-01"
          description: Version of privacy policy accepted
      required: [handle, device, keys, acceptedPrivacyVersion]
      oneOf:
        - required: [password]
          properties:
            passkey:
              not: {}
        - required: [passkey]
          properties:
            password:
              not: {}
    PasswordLoginRequest:
      type: object
      description: |
        Password-based login (fallback for devices without passkey).
        MUST provide EITHER deviceId (existing device) OR newDevice+newDeviceKeys (new device).
      oneOf:
        - title: Existing device login
          type: object
          properties:
            handle:
              type: string
            password:
              type: string
            deviceId:
              type: string
              format: uuid
              description: Existing device ID (for re-login)
          required: [handle, password, deviceId]
        - title: New device login
          type: object
          properties:
            handle:
              type: string
            password:
              type: string
            newDevice:
              $ref: "#/components/schemas/DeviceBootstrap"
              description: New device info (for login from new device)
            newDeviceKeys:
              $ref: "#/components/schemas/InitialKeys"
              description: Keys for new device (required with newDevice)
          required: [handle, password, newDevice, newDeviceKeys]
    PasskeyLoginOptions:
      type: object
      description: WebAuthn options for passkey login (step 1 response)
      properties:
        challenge:
          type: string
          description: Base64URL encoded challenge
        timeout:
          type: integer
          example: 60000
          description: Timeout in milliseconds
        rpId:
          type: string
          example: "ikincikanal.com"
          description: Relying party ID
        allowCredentials:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Base64URL credential ID
              type:
                type: string
                enum: [public-key]
              transports:
                type: array
                items:
                  type: string
                  enum: [usb, nfc, ble, internal, hybrid]
            required: [id, type]
        userVerification:
          type: string
          enum: [required, preferred, discouraged]
          default: preferred
      required: [challenge, timeout, rpId, allowCredentials]
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 3600
        user:
          $ref: "#/components/schemas/User"
        device:
          $ref: "#/components/schemas/Device"
      required: [accessToken, refreshToken, tokenType, expiresIn, user, device]

    # ==================== USER ====================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
          example: fatih
        fullHandle:
          type: string
          example: fatih@ikincikanal.com
          description: Full federated handle
        homeServer:
          type: string
          example: ikincikanal.com
        displayName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, suspended]
        settings:
          $ref: "#/components/schemas/UserSettings"
      required: [id, handle, fullHandle, homeServer, createdAt, status]
    UserSettings:
      type: object
      properties:
        defaultDisappearingSeconds:
          type: integer
          nullable: true
          description: Default disappearing message timer (null = never)
        allowSealedSender:
          type: boolean
          default: true
          description: Accept sealed sender messages
        readReceipts:
          type: boolean
          default: false
          description: Send read receipts (privacy tradeoff)
        typingIndicators:
          type: boolean
          default: false
          description: Send typing indicators (privacy tradeoff)
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        deviceName:
          type: string
          example: "iPhone 15 Pro"
        registrationId:
          type: integer
        identityKeyPub:
          type: string
          description: Base64 Ed25519 public key
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, deviceName, registrationId, identityKeyPub, createdAt, lastActiveAt]
    DeviceBootstrap:
      type: object
      properties:
        deviceName:
          type: string
          maxLength: 64
        registrationId:
          type: integer
          minimum: 1
          maximum: 16380
        identityKeyPub:
          type: string
          description: Base64 Ed25519 public key
      required: [deviceName, registrationId, identityKeyPub]
    AddDeviceRequest:
      type: object
      properties:
        device:
          $ref: "#/components/schemas/DeviceBootstrap"
        keys:
          $ref: "#/components/schemas/InitialKeys"
      required: [device, keys]

    # ==================== KEYS ====================
    InitialKeys:
      type: object
      properties:
        signedPreKey:
          $ref: "#/components/schemas/SignedPreKey"
        oneTimePreKeys:
          type: array
          minItems: 10
          maxItems: 200
          items:
            $ref: "#/components/schemas/OneTimePreKey"
      required: [signedPreKey, oneTimePreKeys]
    SignedPreKey:
      type: object
      properties:
        keyId:
          type: integer
        publicKey:
          type: string
          description: Base64 X25519 public key
        signature:
          type: string
          description: Base64 Ed25519 signature
      required: [keyId, publicKey, signature]
    OneTimePreKey:
      type: object
      properties:
        keyId:
          type: integer
        publicKey:
          type: string
          description: Base64 X25519 public key
      required: [keyId, publicKey]
    UploadKeysRequest:
      type: object
      properties:
        signedPreKey:
          $ref: "#/components/schemas/SignedPreKey"
        oneTimePreKeys:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: "#/components/schemas/OneTimePreKey"
      required: [signedPreKey, oneTimePreKeys]
    PreKeyBundle:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        registrationId:
          type: integer
        identityKeyPub:
          type: string
        signedPreKey:
          $ref: "#/components/schemas/SignedPreKey"
        oneTimePreKey:
          allOf:
            - $ref: "#/components/schemas/OneTimePreKey"
          nullable: true
          description: May be null if exhausted
      required: [userId, deviceId, registrationId, identityKeyPub, signedPreKey]

    # ==================== MESSAGES ====================
...
    PasskeyAttestation:
      type: object
      description: |
        WebAuthn attestation for passkey REGISTRATION.
        Response from navigator.credentials.create()
      properties:
        id:
          type: string
          description: Base64URL credential ID
        rawId:
          type: string
          description: Base64URL raw credential ID
        type:
          type: string
          enum: [public-key]
        response:
          type: object
          properties:
            clientDataJSON:
              type: string
              description: Base64URL encoded
            attestationObject:
              type: string
              description: Base64URL encoded CBOR attestation
            transports:
              type: array
              items:
                type: string
                enum: [usb, nfc, ble, internal, hybrid]
              description: Supported transports for this credential
          required: [clientDataJSON, attestationObject]
        authenticatorAttachment:
          type: string
          enum: [platform, cross-platform]
          description: platform = built-in (Face ID), cross-platform = security key
      required: [id, rawId, type, response]
    PasskeyAssertion:
      type: object
      description: |
        WebAuthn assertion for passkey LOGIN.
        Response from navigator.credentials.get()
      properties:
        id:
          type: string
          description: Base64URL credential ID
        rawId:
          type: string
          description: Base64URL raw credential ID
        type:
          type: string
          enum: [public-key]
        response:
          type: object
          properties:
            clientDataJSON:
              type: string
              description: Base64URL encoded
            authenticatorData:
              type: string
              description: Base64URL encoded authenticator data
            signature:
              type: string
              description: Base64URL encoded signature
            userHandle:
              type: string
              nullable: true
              description: Base64URL encoded user handle (may be null)
          required: [clientDataJSON, authenticatorData, signature]
        authenticatorAttachment:
          type: string
          enum: [platform, cross-platform]
      required: [id, rawId, type, response]

    # ==================== FEDERATION ====================
...